Simple Database Updated
On the 25th of March, I found a better and simple solution for the database. Instead of having Instructor, Admin, and Client tables, I have realized that is better to have one table called App User and a new attribute called Role. This new attribute would help me to offer roles to the user and depending on this each user will have their own privileges.